Moved header to common and use .catch for fetch
