+++ src/API/fetch.jsx
@@ -1,21 +1,8 @@
-const URL = `https://youtube.googleapis.com/youtube/v3/search?key=${process.env.REACT_APP_YOUTUBE}`
 
+const URL = `https://youtube.googleapis.com/youtube/v3/search?key=${process.env.REACT_APP_YOUTUBE}`;
 
+const displayAllVideos = () => {
+  return fetch(`${URL}`).then((response) => response.json());
+};
 
- const displayAllVideos = () => {
-
-    return fetch(`${URL}`)
-        .then(response => response.json())
-    
-
-}
-
-export default displayAllVideos
-
-
-
-
-
-
-
-
+export default displayAllVideos;
