+++ src/Commons/Header.css
@@ -3,20 +3,6 @@
   box-sizing: border-box;
 }
 
-/* header {
-  background-color: #f9f9f9;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  height: 56px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  z-index: 100;
-} */
-
 .hamburger {
   width: 100%;
   height: 5em;
