--- src/Commons/Header.jsx
@@ -1,1 +1,0 @@
-'Hello'
